feat(data): refactored to support async API calls
The store now loads data from the API using Promises to handle async
loading and dispatches. It melted my brain a little, but I added a
metric fuckton of comments to try and make sure it continues to make
sense even after I haven't looked at it for a while.

re #2, #10